index: Fix compiler warning
Problem:
----------------------------
index.c:420:18: warning: comparison of unsigned enum expression < 0 is
always false [-Wtautological-compare]
        if (type < 0 || type >= XATTROP_TYPE_END)
            ~~~~ ^ ~
1 warning generated.
----------------------------
Fix:
Do a type cast before comparing.

BUG: 1335232
Reviewed-on: http://review.gluster.org/14318
Smoke: Gluster Build System <jenkins@build.gluster.com>
NetBSD-regression: NetBSD Build System <jenkins@build.gluster.org>
CentOS-regression: Gluster Build System <jenkins@build.gluster.com>

+++ xlators/features/index/src/index.c
@@ -417,7 +417,7 @@ index_dec_link_count (index_priv_t *priv, index_xattrop_type_t type)
 char*
 index_get_subdir_from_type (index_xattrop_type_t type)
 {
-        if (type < 0 || type >= XATTROP_TYPE_END)
+        if ((int)type < 0 || (int)type >= (int)XATTROP_TYPE_END)
                 return NULL;
         return index_subdirs[type];
 }
